update unstable download

+++ install-cli.sh
@@ -158,14 +158,21 @@ if [ -z "$CLI_RELEASE_TAG" ]; then
 fi
 
 if [ "$1" = "unstable" ]; then
-    git clone https://github.com/iotexproject/iotex-core.git
-    cd iotex-core
-    LATEST_TAG=$(git describe --tags $(git rev-list --tags --max-count=1))
-    git checkout $LATEST_TAG
-    make $INSTALL_NAME
-    sudo mv bin/$INSTALL_NAME $INSTALL_DIRECTORY/
-    cd .. && rm -rf iotex-core
-    exit 0
+    if [ -z "$2" ]; then
+        echo "Usage: sh $0 unstable [branch | tag]\n"
+        echo "eg: sh $0 unstable master or v1.14.0\n"
+        echo "Here is the last 10 tags:\n"
+        echo $(git describe --tags $(git rev-list --tags --max-count=10))
+        exit 1
+    else
+        git clone https://github.com/iotexproject/iotex-core.git
+        cd iotex-core
+        git checkout $2
+        make $INSTALL_NAME
+        sudo mv bin/$INSTALL_NAME $INSTALL_DIRECTORY/
+        cd .. && rm -rf iotex-core
+        exit 0
+    fi
 else
     # fetch the real release data to make sure it exists before we attempt a download
     downloadJSON RELEASE_DATA "$RELEASES_URL/tag/$CLI_RELEASE_TAG"
